enumerate blueprints explicitly in relengapi.app
