+++ relengapi/tests/test_lib_introspection.py
@@ -6,12 +6,6 @@ from nose.tools import eq_
 from relengapi.lib import introspection
 
 
-def test_get_blueprints():
-    blueprints = introspection.get_blueprints()
-    base_blueprint = [bp for bp in blueprints if bp.name == 'base'][0]
-    eq_(base_blueprint.dist.project_name, 'relengapi')
-
-
 def test_get_distributions():
     distributions = introspection.get_distributions()
     eq_(distributions['relengapi'].project_name, 'relengapi')
